use atom with threads as new and loading
